+++ server/sv_ents.c
@@ -577,7 +577,7 @@ void SV_FatPVS (vec3_t org)
 			continue;		// already have the cluster we want
 		src = CM_ClusterPVS(leafs[i]);
 		for (j=0 ; j<longs ; j++)
-			((long *)fatpvs)[j] |= ((long *)src)[j];
+			((int32_t *)fatpvs)[j] |= ((int32_t *)src)[j];
 	}
 }
